Refatorar componente de layout
